Don't gate adding to the frontier
Previously there was a bug in the system in which the frontier couldn't find the syntax error because it had not been introduced to the frontier yet, however it wasn't introducing new blocks anymore.

The frontier already sorts itself so it's fine if we add a new block to the frontier each time. This helps us guarantee that every line is eventually added to the search space. Since an empty string is parsable, it guarantees that we find a solution even if that solution says "the syntax error is somewhere in this entire document.