Expose registering a codeblock to frontier
When a block is added to the frontier, we remove it's lines from the indent_hash so that we do not accidentally re-create the same block multiple times. Previously this behavior was tied only to pushing a block to the frontier. 

This change exposes the behavior so that it can be invoked separately from pushing an object onto the frontier. Also when we generate a new block, we assume that it we don't want to re-generate the exact same block again (the method name is `next_block` after all) so we can go ahead and "register" it so we remove the appropriate lines from indent hash.

This also allows us to generate blocks and evaluate them without having to put them back into the frontier without needing to worry about generating the same block again.