Record code search
While debugging a code search it's very useful to see not just what is returned, but also the process through which that result was created. This commit adds that functionality as well as a test. In the process of writing this code, the internals of the search were refactored a bit also I found a bug where I was accidentally calling `expand_until_neighbors` when I meant to call `expand_until_next_boundry`. This broke a bunch of very brittle test cases