Use indentation code blocks instead of code fences
When copying and pasting the output to GitHub issues I found that I could not put it inside of code blocks since the output itself has code blocks. To alleviate this we can use indentation to mark the code block instead. This allows us to directly copy and paste the output into GitHub.