Refactor: Immutable CodeBlocks
To be able to have CodeBlock class not depend on knowledge of all code lines, we need to move the code expansion logic outside of the code class.

It's now being represented in the BlockExpand class which can be independently tested. To simplify writing this class, a helper class AroundBlockScan was introduced. This class automates some of the chores around searching before/after a given block.

The IndentScan class has been renamed to ParseBlocksFromIndentLine to be a little more descriptive about it's purpose and goal, but I'm still struggling to name it.