Stable sort CodeBlocks
Stable sort means that when elements are entered into an array that have the same value, they will be sorted in the same order. For example:

```
MyValue = Struct.new(:value, :name)

[MyValue.new(1, :a), MyValue.new(1, :b), MyValue.new(1, :c)]
  .sort_by {|x| x.value }
  .map(&:name)
```

You would expect this to always produce `[:a, :b, :c]` since that's the original order that they're entered in, however since ruby's sort is not "stable" this is not the case.

This non-determinism causes tests to produce different results on different systems which leads to CI failures. We can mitigate this by sorting by the value we want first, then by some consistent secondary value. In this case I'm using line number for the first code block.