Fix code context capturing
Previously when we were trying to capture the "neighbors" context, we were using the same logic as when trying to "expand" a block. This logic looks for unmatched keyword pairs, so when it looks up it will stop at `def foo` but not at `def foo; end`. For grabbing context we want to stop after the first contextual pair so we cannot use that same method.

Instead a new method that is specialized to finding context is introduced to the AroundBlockScan class.

 When trying to capture code context, we don't want to look for a dangling keyword, we want the first matched pair