Fix perf issue with combination
After dead_end has determined that the problem has been found, it then goes back through all it's code blocks to determine the smallest set of blocks which hold all the errors (since there can be multiple errors and multiple blocks could hold the same error).

It does this by building a "combination" of all possible block results and then checking each. To see, look at this spec:

```ruby
      expect(
        CodeFrontier.combination([:a, :b, :c, :d])
      ).to eq(
        [
          [:a],[:b],[:c],[:d],
          [:a, :b],
          [:a, :c],
          [:a, :d],
          [:b, :c],
          [:b, :d],
          [:c, :d],
          [:a, :b, :c],
          [:a, :b, :d],
          [:a, :c, :d],
          [:b, :c, :d],
          [:a, :b, :c, :d]
        ]
      )
``` 

Imagine that :a, :b, :c, and :d are each code blocks. If we have 4 blocks in our frontier it generates 15 possibilities to evaluate.

- 4 => 15
- 5 => 31
- 6 => 63
- 7 => 127

In this failing scenario there are a large number of blocks 23 blocks to be exact.

- 23 => 8,388,607

It takes 4.7 seconds just to generate all these combinations with numeric values, which is not sustainable. This doesn't even include the time to perform the parse check 8 million times, just to generate 8 million arrays with all possible combinations. YIKEs.

This wasn't previously as issue because as code blocks "expand" they consolidate when a larger block contains smaller blocks they get grouped together.

The reason this example is so "bad", is that there's a lot of blocks all at the same indentation level, so they've not had a chance to consolidate yet.

The fix is to evaluate fewer combinations. We know that removing valid code from an invalid document cannot create a valid document, so we can remove all blocks that are already valid.

When we've filtered to only invalid blocks, there's only 1 in the frontier which comes back almost instantaneously. This patch reduces the time to find the syntax error from about 9 seconds to 0.03 seconds.

Not bad