More efficient frontier unvisited line structure
Previously we were keeping track of the lines that were not in the frontier via an @indent_hash with the keys as indentation and value as an array.

This was inefficient and caused us to constantly be sorting the keys. Instead we can get the same information by only sorting once.

As a bonus perf optimization CodeLine#indent_index generates a key for stable sort and memoizes it so it only has to be allocated once per line.