[close #32] Improve trailing end results
When the last keyword in a given logical "block" is missing the `end` then it's unclear if the missing end came from the surrounding block, or from the last keyword. Example:

```ruby
class Dog
  def bark
    puts "woof"
end
```

Before this patch the output of this code will be:

```
    ❯ 1  class Dog
``

With this patch the output of this code will be:

```
    ❯ 1  class Dog
    ❯ 2    def bark
    ❯ 4  end
```

There's some key realizations to get to this point. First, while the output of the "before" case only looks like it's one line, it's actually a much larger block that holds down to line 4 but everything else is hidden. Second: We can detect for this scenario using lexing but starting at the last end of the problem block and working backwards to find a mismatched keyword.