[close #42] Annotate NoMethodError
When a NoMethodError is raised in a non-production environment, instead of producing:

```
scratch.rb:7:in `call': undefined method `upcase' for nil:NilClass (NoMethodError)
```

Ruby programs with `dead_end` required will produce:

```
scratch.rb:7:in `call': undefined method `upcase' for nil:NilClass (NoMethodError)


  1  class Pet
  6    def call
❯ 7      puts "Come here #{@neam.upcase}"
  8    end
  9  end

```

This behavior is disabled in "production" environments as it requires reading from disk and parsing. The main behavior of DeadEnd is reasonably production safe as syntax errors stop a program from executing further and are limited to the number of files on disk. A NoMethodError in production is likely to be hit repeatedly.