Display raw errors from Ruby's parser
Currently we're trying to make sense of the errors by the whole document and delivering one "banner" stating the cause of the problem.

This is based on the "who dis" ripper subclass which only captures one error.

With this move, we're going to multi errors per output and not attempting to re-explain the existing ripper error. The main focus is showing the problem rather than explaining it.

Example:

```
$ ./exe/dead_end spec/fixtures/rexe.rb.txt
--> /Users/rschneeman/Documents/projects/dead_end/spec/fixtures/rexe.rb.txt

syntax error, unexpected `def', expecting ')'
syntax error, unexpected `end', expecting end-of-input

   16  class Rexe
❯  40    class Options < Struct.new(
❯  41        :input_filespec,
❯  42        :input_format,
❯  43        :input_mode,
❯  44        :loads,
❯  45        :output_format,
❯  46        :output_format_tty,
❯  47        :output_format_block,
❯  48        :requires,
❯  49        :log_format,
❯  71    end
  551  end

syntax error, unexpected end-of-input, expecting `end'

   16  class Rexe
   40    class Options < Struct.new(
   71    end
❯  77    class Lookups
❯ 140      def format_requires
❯ 148    end
  551  end
```

Eventually We can re-work this logic to attempt to do a better job at explaining errors based on lex values in each captured block rather than using regex values on the parser errors.