Remove WhoDisSyntaxError
This class was used for detecting the syntax error type and emitting a "better" error than the parser. It made several incorrect assumptions, mainly that the document would only have a single syntax error.

Over time this has proven to be an ineffective if not outright confusing annotation. With all the usage removed, these classes can also be removed.

Ideally we will come back and add a better, more robust way to improve the errors, at which time we can revisit the banner code which will continue to live on his git history.