Lex based banner
Described in https://github.com/zombocom/dead_end/issues/68#issuecomment-955575369. This implements a method of checking the source code for logically missing component pairs:

- {}
- []
- |<value>|
- ()
- keyword/end

It seems to be much more robust than using a regex on the parser errors. It also seems to provide more confident results than the parser errors.

For cases where we are not confident about why code is invalid, we still fall back on the parser error messages. That allows this functionality to act as a progressive enhancement rather than a wholesale replacement.

We may be able to specialize other checks in the future as well. The most common one for me that comes up is missing a trailing comma in a hash/array/method-call:

```
    query = Cutlass::FunctionQuery.new(
      port: port # <== here
      body: body
    ).call
```