Fix incorrect code removal
As listed in #95 there are other cases where we have valid code nested in a way we didn't previously handle (via `CleanDocument`). The specific case is:

```ruby
def ruby_install_binstub_path(ruby_layer_path = ".")
  @ruby_install_binstub_path ||=
    if ruby_version.build?
      "#{build_ruby_path}/bin"
    elsif ruby_version
      "#{ruby_layer_path}/#{slug_vendor_ruby}/bin"
    else
      ""
    end
end
```

As it's being through being validated/hidden, the inside is removed first, which generates an invalid code block.

This fix is to re-check all code in a block even if it is "hidden".

This conveniently drastically improves the results of #88 beyond just that small case, the whole output becomes very focused

close #95
close #88