Faster command line spec
This is the slowest test in the suite, taking longer than 1 second. I had the logic duplicated for some weird reason, removing that got me twice as fast. Since we need to boot 2 processes (the slow process), we can spawn them both before waiting on the first one to finish.