Fix double output
When we proxy `Object.load` to `Kernel.load` we were firing `DeadEnd.handle_error` twice. Switching the test over to a `load` instead of a `require_relative` makes the error show up, which can then be asserted via adding the `once` call in rspec.

I still don't know why this code was needed in the first place.

close #97