Insertion sort
Before:  0.230749   0.005489   0.236238 (  0.237043)
After:  0.197075   0.005009   0.202084 (  0.202950)


Using `qcachegrind` on Mac, generating an output with `RubyProf::CallTreePrinter` I saw:


![](https://www.dropbox.com/s/xian4mbsgvi8xr7/Screen%20Shot%202021-11-03%20at%203.00.05%20PM.png?raw=1)

That a lot of time was being spent in `CodeFrontier#<<`, specifically in in the `sort!` function and `CodeBlock<=>`.

Because we control insertion into the array, we know that it is always sorted. We can leverage this info to place new elements in at the right location instead of placing them and then re-sorting the whole array just to place one element.

I experimented with iterating from front to back, and in reverse. I found that in my test case reverse took 7,373 comparisons while forwards took 3,130 however both represent savings.

After changing this logic:

![](https://www.dropbox.com/s/oba4mmrmlndvkac/Screen%20Shot%202021-11-03%20at%202.58.40%20PM.png?raw=1)

You can see `sort!` no longer shows up. Instead we're seeing `reject!` as a hotspot (though only taking up 8.4% time instead of 54% time seen in sort). 

Strangely such a large bump in results only yielded ~1.15x faster overall performance change. It's still worth it, but not in line with what I expected from the tools.