Faster CleanDocument
Benchmarking with `ruby_buildpack.rb.txt`:


Before:   0.195756   0.005201   0.200957 (  0.202021)
After:   0.148460   0.003634   0.152094 (  0.152692)

I was using ruby prof to try to find perf opportunities when I saw this output from the `CallStackPrinter`:

![](https://www.dropbox.com/s/7m6wi502fqoel2g/Screen%20Shot%202021-11-03%20at%204.03.10%20PM.png?raw=1)

I noticed a lot of time was being spent in CleanDocument which was curious as it's an N=1 process that happens before we do any looping or iteration. 

I also saw output from the `RubyProf::GraphHtmlPrinter` that we were creating many many LexValue objects.

The way CleanDocument works to remove lines with a comment, it first runs lex over the document, and then it uses the lex output to determine which lines have comments that it can remove. 

Once the comments are removed the lexer is run again as you can produce different results with comments removed (specifically :on_ignored_nl). 

That approach works, but is brute force. It means we must lex the document twice which is expensive and generates many LexValue objects. 

We can remove lines based on regex, one thing we can't distinguish via only regex is if someone is in a heredoc or not. So if a line looks like it could be a heredoc string embed like:

```
source <<~EOM
  #{here}
EOM
```

Then we don't remove it as it's very unlikely someone will create a comment with a `#{` pattern in a place we care about (i.e. in-between method calls) like:

```
User
  #{ this would be weird
  .first
```

With this approach we reduce lexing passes from N=2 to N=1.

As a side note, this didn't really show up as a hotspot in the perf tools. Even looking back at the screenshots here, it's not obvious that there was a problem here or I could get a 1.3x speed boost by making this change. Even running it before and after, it's not clear it got faster based on the output:


![](https://www.dropbox.com/s/j8t7xd0i5unyim3/Screen%20Shot%202021-11-03%20at%204.11.55%20PM.png?raw=1)

(Before is top, after is bottom)


I guess you can see before was  1.1264283200034697 while after was 0.8452127130003646 at the top. It's not clear why CleanDocument is showing 17% after when it was only showing 12% before

Side by side it maybe makes a bit more sense in that the problem dropped in relative priority down the list, however I'm still not entirely clear what the numbers are indicating:

(Before left, After, right)

![](https://www.dropbox.com/s/jgxitymh2wtoxpr/Screen%20Shot%202021-11-03%20at%204.30.08%20PM.png?raw=1)

You can see the results for yourself https://www.dropbox.com/sh/kueagukss7ho2rm/AABk3V8FQXaFWugKI21ADZJ0a?dl=0