Failing test for aspirational perf improvements
This file is 9k lines long and even with the currently optimized algorithm takes about 1.6s to search. The goal is to find a way to get it under 1 second.

If someone is working with such large files they can set their DEAD_END_TIMEOUT=2 and be perfectly fine, but perf improvements that would allow this file to parse under 1 second would hopefully also speed up other cases.