Totally new tree search
Experimental work for a "tree" search described in https://github.com/zombocom/dead_end/issues/118#issuecomment-1004950659
(I'm using "tree" to describe the "pyramid" concept in that post).

The short version of the concept is this: Recursively build a tree structure that is sorted in such a way
that I can call `bsearch` to reduce/eliminate calls to Ripper.

To do this I recursively build an individual "tree" which **should** break a document up into a large
logical code block. Then I repeat the process for the code above and below that tree.

Once done I can iteratively determine which trees hold syntax errors to focus my search. Then I can
use the tree structure to isolate the indentation level (can use bsearch and Ripper) as well as find and split
an invalid code block from that indentation level (also using bsearch and Ripper). In all it reduces time to
find syntax errors on the 9k long line file from 1.6s to 0.61s (roughly in a third).

The biggest challenge is that the old algorithm essentially had some redundency if the blocks are not created
correctly. This algorithm is much more sensitive and I've not found a way to back out of a bad/incorrect search
plan yet, though that may come with time.


This commit breaks a lot of existing tests, it's mostly meant as a trial WIP to validate different concepts.