Add interface to require dead_end without core_ext (#119)
* Add interface to require dead_end without core_ext

Currently `dead_end` works by monkey patching `require` and friends.

If someone wants to programmatically execute dead_end manually via `DeadEnd.handle_error` without monkey patches they can now do that by setting the environment variable `DISABLE_DEAD_END_CORE_EXT=1`.

* Move api to `dead_end/api` instead of env var

Because setting environment variables on systems may be difficult or inconsistent we are moving the API to be require based. Now anyone who wants to use dead_end without mutating `require` can do so via requiring `dead_end/api`.

* Update lib/dead_end/api.rb

Co-authored-by: Jake Zimmerman <zimmerman.jake@gmail.com>

* Add explicit check for Syntax Error

As suggested in https://github.com/zombocom/dead_end/pull/119#discussion_r753592117 we are now explicitly checking the type of the incoming error. Rather than raise a new error, I've chosen to emit a warning and re-raise the original error. This is a similar pattern to the case where we are not able to detect the filename from the error message. The goal is to provide visibility into the source of the behavior, but to break as few expectations as possible. 

The `DeadEnd.handle_error` should (ideally) never raise an error to the end user. The purpose of the library is to provide progressive error enhancement. If it fails it means we cannot make a better error, so therefore we should warn (so the user can fix or report the problem) and then fall back to the original error.

I also added a test for the failure error condition where the filename cannot be pulled out of the SyntaxError message which was added in https://github.com/zombocom/dead_end/pull/114.

Co-authored-by: Jake Zimmerman <zimmerman.jake@gmail.com>