WIPWIPWIP
Built a tree, now need to figure out how to search it

- It gives me large blocks that I can use to narrow search.

If we remove all of the inner valid blocks, then we're left with the highest indentation invalid block. How to do recursively?

parents.first.last and parents.last.first tends to give a balanced block (makes sense)

Maybe we recurse until we find a block who's parent.first.last and parents.last.first is valid, but it is not valid