WIP block expand
0 failing tests, one commented out in block expand spec though

Before:

```
41.08% 	0.11% 	2.13 	0.01 	0.00 	2.12 	2291 DeadEnd::CodeSearch#expand_existing 
```

After:

```
38.96% 	0.10% 	1.77 	0.00 	0.00 	1.76 	2031 DeadEnd::CodeSearch#expand_existing 
```

A roughly 2% speed increase.


Reduces overall iterations from 2301 to 2041 (12% fewer iterations), but does not increase speed by 12% either due to these being fast operations that we're eliminating or the extra cost of generating the blocks is not cheap.

We can check these assumptions. Looking at new operations added (that may hold optimizations):

```
0.79% 	0.02% 	0.04 	0.00 	0.00 	0.04 	1253 DeadEnd::AroundBlockScan#captured_meaningful? 
```

We could likely optimize this to capture this information at scan time instead of having to allocate an array and iterate over it. However that would at best be a 0.79% speedup. Maybe worth doing, but not top priority.

And

```
0.90% 	0.21% 	0.04 	0.01 	0.00 	0.03 	4993 DeadEnd::AroundBlockScan#captured_current_indent? 
```

This method is already quite optimized. If we made captured_meaningful constant time, we could perhaps swap the order of our checks and avoid calling this some times to bring runtime down. Best guess would be reducing calls by 25% which would bring us to a 0.225% speedup which is worth exploring, but is absolutely not a priority. 


Overall:

Before

```
3.31% 	0.13% 	0.17 	0.01 	0.00 	0.16 	2291 DeadEnd::BlockExpand#expand_neighbors
```

After:

```
3.13% 	0.12% 	0.14 	0.01 	0.00 	0.14 	2031 DeadEnd::BlockExpand#scan_current_indent 
```


So it's not significantly slower. One other thought is that we've optimized the old code already so we may be performing checks that aren't needed (or as needed). Specifically the CodeBlock#valid? Checks to see if all entries are empty and if they are it skips ripper.

On main this occurs 336 times versus on this branch 76 times (while total uncached calls to this method are 3361). We pay for this check for every uncached call, but only use it Before (336/3361 # => 9% hit rate) After (76/3361 # => 2% hit rate). 

Getting rid of this extra check yields tiny speed improvements:

Before

```
9.18% 	0.19% 	0.42 	0.01 	0.00 	0.41 	6524 DeadEnd::CodeBlock#valid?
```

After

```
8.75% 	0.15% 	0.40 	0.01 	0.00 	0.39 	6524 DeadEnd::CodeBlock#valid?
```

For a 0.43% speed up. Assuming we've realized 2/3 of all possible gains for removing this extra check (336-76)/336, if we captured 100% of those cases and never called valid we would expect to only see a total of 0.645% speedup which is quite negligible (So it's likely not worth going to extreme lengths to capture those extra cases).


Reducing the number of empty blocks by 76 would bring the overall iteration count down from 2041 to 1965 for a reduction of ~3%.


Making the optimization of removing the empty line check in CodeBlock#valid? Is somewhat of a wash. I think we're probably within margin of error of our ability to measure. While that method seems to be faster, the numbers fluctuate and overall it doesn't seem to change overall time spent to CodeSearch#call